0x02. C - Functions, nested loops

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -9,7 +9,7 @@
 int main(void)
 {
 	unsigned long fib1 = 0, fib2 = 1, sum;
-	float total_sum;
+	int total_sum;
 
 	while (1)
 	{
@@ -23,7 +23,7 @@ int main(void)
 		fib1 = fib2;
 		fib2 = sum;
 	}
-	printf("%0f\n", total_sum);
+	printf("%d\n", total_sum);
 
 	return (0);
 }
